% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_conversions.R
\name{ac_to_ha}
\alias{ac_to_ha}
\title{ac_to_ha}
\usage{
ac_to_ha(dat_ac)
}
\arguments{
\item{dat_ac}{A vector of values in acres}
}
\value{
A vector of values in hectares; multiplied by ~0.41
}
\description{
ac_to_ha
}
