% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_conversions.R
\name{ha_to_ac}
\alias{ha_to_ac}
\title{ha_to_ac}
\usage{
ha_to_ac(dat_ha)
}
\arguments{
\item{dat_ha}{A vector of values in hectares}
}
\value{
A vector of values in acres; multiplied by ~2.47
}
\description{
ha_to_ac
}
