% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_conversions.R
\name{ha_to_m2}
\alias{ha_to_m2}
\title{ha_to_m2}
\usage{
ha_to_m2(dat_ha)
}
\arguments{
\item{dat_ha}{A vector of values in hectares}
}
\value{
A vector of values in m2; multiplied by 10 000
}
\description{
ha_to_m2
}
