% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_conversions.R
\name{kg_to_lb}
\alias{kg_to_lb}
\title{kg_to_lb}
\usage{
kg_to_lb(dat_kg)
}
\arguments{
\item{dat_kg}{A vector of values in kg}
}
\value{
A vector of values in pounds; multiplied by ~0.45
}
\description{
kg_to_lb
}
