% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_conversions.R
\name{lb_to_kg}
\alias{lb_to_kg}
\title{lb_to_kg}
\usage{
lb_to_kg(dat_lb)
}
\arguments{
\item{dat_lb}{A vector of values in pounds}
}
\value{
A vector of values in kilograms; multiplied by 2.204624
}
\description{
lb_to_kg
}
