% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrs_conversions.R
\name{mm_to_in}
\alias{mm_to_in}
\title{mm_to_in}
\usage{
mm_to_in(dat_mm)
}
\arguments{
\item{dat_mm}{A vector of values in mm}
}
\value{
A vector of values in inches; multiplied by ~25.4
}
\description{
mm_to_in
}
